+++ libs/providers/langchain-baseten/CHANGELOG.md
@@ -1,5 +1,5 @@
 # @langchain/baseten
 
-## 0.0.1
+## 0.1.0
 
 Initial release of the standalone Baseten provider package for LangChain.js.
