+++ libs/providers/langchain-baseten/src/index.ts
@@ -1,3 +1,3 @@
 export { ChatBaseten, normalizeToolCallChunks } from "./chat_models.js";
-export { normalizeModelUrl } from "./types.js";
+export { normalizeModelUrl } from "./converters/params.js";
 export type { BasetenChatInput } from "./types.js";
