+++ libs/providers/langchain-baseten/src/types.ts
@@ -1,47 +1,6 @@
 import type { ChatOpenAIFields } from "@langchain/openai";
 
 /**
- * Default base URL for Baseten's managed inference API.
- * Supports all open-source models hosted on Baseten's Model APIs.
- *
- * For self-deployed models, override with the model-specific URL:
- * `https://model-{model_id}.api.baseten.co/v1`
- */
-export const DEFAULT_BASE_URL = "https://inference.baseten.co/v1";
-
-/**
- * Default environment variable name for the Baseten API key.
- */
-export const DEFAULT_API_KEY_ENV_VAR = "BASETEN_API_KEY";
-
-/**
- * Normalize a dedicated model URL to OpenAI-compatible `/sync/v1` format.
- *
- * Baseten dedicated model endpoints come in several forms:
- * - `.../predict` -> converted to `.../sync/v1`
- * - `.../sync` -> appended with `/v1`
- * - anything else -> ensures trailing `/v1`
- *
- * See: Python `langchain-baseten._normalize_model_url`
- */
-export function normalizeModelUrl(url: string): string {
-  if (url.endsWith("/predict")) {
-    return `${url.slice(0, -"/predict".length)}/sync/v1`;
-  }
-  if (url.endsWith("/sync")) {
-    return `${url}/v1`;
-  }
-  if (!url.endsWith("/v1")) {
-    let trimmed = url;
-    while (trimmed.endsWith("/")) {
-      trimmed = trimmed.slice(0, -1);
-    }
-    return `${trimmed}/v1`;
-  }
-  return url;
-}
-
-/**
  * Input fields for constructing a `ChatBaseten` instance.
  *
  * Extends `ChatOpenAIFields` since Baseten exposes an OpenAI-compatible API.
